revert to switching between env var and gradle property
it's much more obvious in the ci script to use a short
variable name vs. the ORG_GRADLE_PROJECT prefix required
to make it appear as a gradle property.

additionally, the property will likely be used by Javascript
builds and would require the 'long' version to be used
by tooling.